pyproject.toml
[project.urls]
"Homepage" = "https://github.com/liamwazherealso/news-ml-scraper"
"Bug Tracker" = "https://github.com/liamwazherealso/news-ml-scraper/issues"

[tool.poetry]
name = "news-ml-scraper"
version = "0.0.1"
description = "Scraper for News ML project"
authors = ["Liam Pieri <liamwazherealso@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "news_ml_scraper"}]

[tool.poetry.dependencies]
python = "^3.11"
newspaper3k = "^0.2.8"
boto3 = "^1.26.97"
gnews = "^0.2.7"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.0"
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
